requirements.txt
python-whois
dnspython
requests
streamlit
streamlit-extras
jellyfish
numpy